[SpeedDial] Move from lab to core
