+++ docs/src/pages.js
@@ -56,6 +56,7 @@ const pages = [
           { pathname: '/components/drawers' },
           { pathname: '/components/links' },
           { pathname: '/components/menus' },
+          { pathname: '/components/speed-dial' },
           { pathname: '/components/steppers' },
           { pathname: '/components/tabs' },
         ],
@@ -124,7 +125,6 @@ const pages = [
           { pathname: '/components/skeleton' },
           { pathname: '/components/rating' },
           { pathname: '/components/slider-styled' },
-          { pathname: '/components/speed-dial' },
           { pathname: '/components/timeline' },
           { pathname: '/components/toggle-button' },
           { pathname: '/components/trap-focus' },
