+++ packages/material-ui-lab/src/themeAugmentation/overrides.d.ts
@@ -2,9 +2,6 @@ import { AutocompleteClassKey } from '../Autocomplete';
 import { AvatarGroupClassKey } from '../AvatarGroup';
 import { PaginationClassKey } from '../Pagination';
 import { PaginationItemClassKey } from '../PaginationItem';
-import { SpeedDialClassKey } from '../SpeedDial';
-import { SpeedDialActionClassKey } from '../SpeedDialAction';
-import { SpeedDialIconClassKey } from '../SpeedDialIcon';
 import { TabListClassKey } from '../TabList';
 import { TabPanelClassKey } from '../TabPanel';
 import { TimelineClassKey } from '../Timeline';
@@ -24,9 +21,6 @@ export interface LabComponentNameToClassKey {
   MuiAvatarGroup: AvatarGroupClassKey;
   MuiPagination: PaginationClassKey;
   MuiPaginationItem: PaginationItemClassKey;
-  MuiSpeedDial: SpeedDialClassKey;
-  MuiSpeedDialAction: SpeedDialActionClassKey;
-  MuiSpeedDialIcon: SpeedDialIconClassKey;
   MuiTabList: TabListClassKey;
   MuiTabPanel: TabPanelClassKey;
   MuiTimeline: TimelineClassKey;
