+++ packages/material-ui-lab/src/themeAugmentation/props.d.ts
@@ -2,9 +2,6 @@ import { AutocompleteProps } from '../Autocomplete';
 import { AvatarGroupProps } from '../AvatarGroup';
 import { PaginationProps } from '../Pagination';
 import { PaginationItemProps } from '../PaginationItem';
-import { SpeedDialProps } from '../SpeedDial';
-import { SpeedDialActionProps } from '../SpeedDialAction';
-import { SpeedDialIconProps } from '../SpeedDialIcon';
 import { TabListProps } from '../TabList';
 import { TabPanelProps } from '../TabPanel';
 import { TimelineProps } from '../Timeline';
@@ -24,9 +21,6 @@ export interface LabComponentsPropsList {
   MuiAvatarGroup: AvatarGroupProps;
   MuiPagination: PaginationProps;
   MuiPaginationItem: PaginationItemProps;
-  MuiSpeedDial: SpeedDialProps;
-  MuiSpeedDialAction: SpeedDialActionProps;
-  MuiSpeedDialIcon: SpeedDialIconProps;
   MuiTabList: TabListProps;
   MuiTabPanel: TabPanelProps;
   MuiTimeline: TimelineProps;
