+++ packages/material-ui/src/SpeedDial/index.d.ts
@@ -1,0 +1,2 @@
+export { default } from './SpeedDial';
+export * from './SpeedDial';
