+++ packages/material-ui/src/SpeedDial/index.js
@@ -1,0 +1,1 @@
+export { default } from './SpeedDial';
