+++ packages/material-ui/src/SpeedDialAction/index.d.ts
@@ -1,0 +1,2 @@
+export { default } from './SpeedDialAction';
+export * from './SpeedDialAction';
