+++ packages/material-ui/src/SpeedDialAction/index.js
@@ -1,0 +1,1 @@
+export { default } from './SpeedDialAction';
