+++ packages/material-ui/src/SpeedDialIcon/index.d.ts
@@ -1,0 +1,2 @@
+export { default } from './SpeedDialIcon';
+export * from './SpeedDialIcon';
