+++ packages/material-ui/src/SpeedDialIcon/index.js
@@ -1,0 +1,1 @@
+export { default } from './SpeedDialIcon';
