+++ packages/material-ui/src/styles/overrides.d.ts
@@ -74,6 +74,9 @@ import { SkeletonClassKey } from '../Skeleton';
 import { SliderClassKey } from '../Slider';
 import { SnackbarClassKey } from '../Snackbar';
 import { SnackbarContentClassKey } from '../SnackbarContent';
+import { SpeedDialClassKey } from '../SpeedDial';
+import { SpeedDialActionClassKey } from '../SpeedDialAction';
+import { SpeedDialIconClassKey } from '../SpeedDialIcon';
 import { StepButtonClasskey } from '../StepButton';
 import { StepClasskey } from '../Step';
 import { StepConnectorClasskey } from '../StepConnector';
@@ -190,6 +193,9 @@ export interface ComponentNameToClassKey {
   MuiSlider: SliderClassKey;
   MuiSnackbar: SnackbarClassKey;
   MuiSnackbarContent: SnackbarContentClassKey;
+  MuiSpeedDial: SpeedDialClassKey;
+  MuiSpeedDialAction: SpeedDialActionClassKey;
+  MuiSpeedDialIcon: SpeedDialIconClassKey;
   MuiStep: StepClasskey;
   MuiStepButton: StepButtonClasskey;
   MuiStepConnector: StepConnectorClasskey;
