+++ packages/material-ui/src/styles/props.d.ts
@@ -78,6 +78,9 @@ import { SkeletonProps } from '../Skeleton';
 import { SliderProps } from '../Slider';
 import { SnackbarContentProps } from '../SnackbarContent';
 import { SnackbarProps } from '../Snackbar';
+import { SpeedDialProps } from '../SpeedDial';
+import { SpeedDialActionProps } from '../SpeedDialAction';
+import { SpeedDialIconProps } from '../SpeedDialIcon';
 import { StepButtonProps } from '../StepButton';
 import { StepConnectorProps } from '../StepConnector';
 import { StepContentProps } from '../StepContent';
@@ -190,6 +193,9 @@ export interface ComponentsPropsList {
   MuiSlider: SliderProps;
   MuiSnackbar: SnackbarProps;
   MuiSnackbarContent: SnackbarContentProps;
+  MuiSpeedDial: SpeedDialProps;
+  MuiSpeedDialAction: SpeedDialActionProps;
+  MuiSpeedDialIcon: SpeedDialIconProps;
   MuiStep: StepProps;
   MuiStepButton: StepButtonProps;
   MuiStepConnector: StepConnectorProps;
